Added package for the entity fields, separated for more readability (phase, header, comments, and soon fields) + some small re-factoring
